consumers/requirements.txt
kafka-python
psycopg2-binary
clickhouse-driver